+++ catalyst/metrics/_hitrate.py
@@ -8,12 +8,15 @@ from catalyst.metrics.functional._hitrate import hitrate
 
 
 class HitrateMetric(ICallbackBatchMetric):
-    """
-    Calculate the hitrate
-    score given model outputs and targets
-    The precision metric summarizes the fraction of relevant items
+    """Calculates the hitrate.
 
-    Compute mean value of hitrate and it's approximate std value
+    Args:
+        topk_args: list of `topk` for hitrate@topk computing
+        compute_on_call: if True, computes and returns metric value during metric call
+        prefix: metric prefix
+        suffix: metric suffix
+
+    Compute mean value of hitrate and it's approximate std value.
     """
 
     def __init__(
@@ -23,15 +26,7 @@ class HitrateMetric(ICallbackBatchMetric):
         prefix: str = None,
         suffix: str = None,
     ):
-        """
-        Init HitrateMetric
-
-        Args:
-            topk_args: list of `topk` for hitrate@topk computing
-            compute_on_call: if True, computes and returns metric value during metric call
-            prefix: metric prefix
-            suffix: metric suffix
-        """
+        """Init HitrateMetric"""
         super().__init__(compute_on_call=compute_on_call, prefix=prefix, suffix=suffix)
         self.metric_name_mean = f"{self.prefix}hitrate{self.suffix}"
         self.metric_name_std = f"{self.prefix}hitrate{self.suffix}/std"
