+++ catalyst/runners/__init__.py
@@ -1,7 +1,8 @@
 # flake8: noqa
 
-from catalyst.runners.runner import Runner
-from catalyst.runners.supervised import ISupervisedRunner, SupervisedRunner
+
+from catalyst.runners.supervised import ISupervisedRunner
+from catalyst.runners.runner import Runner, SupervisedRunner
 from catalyst.runners.config import ConfigRunner, SupervisedConfigRunner
 
 from catalyst.settings import SETTINGS
