+++ catalyst/runners/hydra.py
@@ -13,8 +13,8 @@ from catalyst.callbacks import CheckpointCallback, ICheckpointCallback
 from catalyst.callbacks.batch_overfit import BatchOverfitCallback
 from catalyst.callbacks.misc import CheckRunCallback, TimerCallback, TqdmCallback
 from catalyst.core import Callback
-from catalyst.core.functional import callback_isinstance
 from catalyst.core.logger import ILogger
+from catalyst.core.misc import callback_isinstance
 from catalyst.core.runner import IRunner
 from catalyst.core.trial import ITrial
 from catalyst.engines import IEngine
@@ -351,7 +351,15 @@ class HydraRunner(IRunner):
 
 
 class SupervisedHydraRunner(ISupervisedRunner, HydraRunner):
-    """@TODO: docs."""
+    """@TODO: docs.
+
+    Args:
+        cfg:
+        input_key:
+        output_key:
+        target_key:
+        loss_key:
+    """
 
     def __init__(
         self,
@@ -361,7 +369,7 @@ class SupervisedHydraRunner(ISupervisedRunner, HydraRunner):
         target_key: str = "targets",
         loss_key: str = "loss",
     ):
-        """@TODO: docs."""
+        """Init."""
         ISupervisedRunner.__init__(
             self,
             input_key=input_key,
